Allow using Botan for the TPM2 TSS' crypto needs
This adds an implementation of the tpm2-tss crypto callbacks. If
enabled, Botan will be used for the client-side crypto functions to
communicate with the TPM. This lets applications shed a transitive
dependency on another crypto library (like OpenSSL or mbedTLS).

The crypto callbacks are available in tpm2-tss 4.0 and later. Before
that, calling TPM2::Context::use_botan_crypto_backend() will result in
an exception.

Co-Authored-By: René Meusel <rene.meusel@rohde-schwarz.com>