Backport: Reject ECDH agreement when peer sent the identity
See also: https://github.com/randombit/botan/pull/4794

BSI's TR-03111 Section 4.3.1 require that we check the resulting point
of the multiplication to not be the point at infinity. However, since
we ensure that our ECC private scalar can never be zero, checking the
peer's input point is equivalent.

Note that the check for infinity in EC_PublicKey is not hit here, as
we decode an EC_AffinePoint directly without creating an ECDH_PublicKey
object.

Co-Authored-By: Amos Treiber <amos.treiber@rohde-schwarz.com>